changed to full precision

+++ qwen-vl-finetune/qwenvl/train/train_qwen.py
@@ -112,7 +112,7 @@ def train(attn_implementation="flash_attention_2"):
             model_args.model_name_or_path,
             cache_dir=training_args.cache_dir,
             attn_implementation=attn_implementation,
-            torch_dtype=torch.float16, #(torch.bfloat16 if training_args.bf16 else None),
+            torch_dtype=torch.float32, #(torch.bfloat16 if training_args.bf16 else None),
         )
         data_args.image_processor = AutoProcessor.from_pretrained(
             model_args.model_name_or_path,
@@ -125,7 +125,7 @@ def train(attn_implementation="flash_attention_2"):
             model_args.model_name_or_path,
             cache_dir=training_args.cache_dir,
             attn_implementation=attn_implementation,
-            torch_dtype=torch.float16#(torch.bfloat16 if training_args.bf16 else None),
+            torch_dtype=torch.float32#(torch.bfloat16 if training_args.bf16 else None),
         )
         data_args.image_processor = Qwen2VLImageProcessor.from_pretrained(
             model_args.model_name_or_path,
